requirements/base.txt
tornado>=5.1.1
pymongo>=3.7.2
requests==2.20.0
psutil


casbin
pycrypto
openpyxl
pyyaml

schedule
paho-mqtt
watchdog